generator/requirements.txt
pyyaml>=6.0
python-osc>=1.8.0
